Fixed indexing bug in python/templates.time_frequency_boudaries
Also significantly rewrote this function to isolate the waveform-
family-dependent parts.  I compromised on generality and usefulness,
leaving the calculation of flow/fhigh to the user and leaving only
one line of code that depends on the waveform family
(the spawaveform.chirptime function).  Other waveform families can
be support by swapping out for this function.