lal_whiten:  more progress towards gap support
This should cause gaps to be treated correctly on input:  they are
recognized as intervals of 0s, the PSD is not updated using FFT blocks
contaminated with 0s, and when an all-0 FFT block is encountered most of
the computations get skipped.  The element is not yet smart enough to
recognize when its output is all zeros, reduce itself to a no-op, and flag
its as a gap.  That will follow.