add segment-list property
This gives the framesrc the ability to produce a continuous stream of data
from disjoint segments.  The framesrc is provided with a single frame cache
as usual, and a seek is performed to the desired start time of the stream
as usual.  If a segment list is provided, only data from the intervals in
the segment list will be read from disk and pushed downstream.  The
intervals outside the segment list will be pushed down stream as gaps.  If
no segment list is provided, all data is read from disk and pushed down
stream.

When using the segment list feature, the gap/non-gap transitions might
result in short buffers being produced, that is buffers smaller than block
size will be produced in order to align their boundaries with the segment
boundaries.  No buffers larger than blocksize will be produced.