lloidparts:  run autochisq in same thread as trigger generator
the trigger generator is essentially free compared compared to the chisq
element and is slaved to the output of the chisq element buffer-by-buffer,
so there's not much point in trying to run them in parallel.  this patch
removes the large queue between the autochisq and trigger generator
elements and places it upstream of the autochisq, between it and the tee
following the togglecomplex.