lal_autochisq:  forgot an adapter flush
when the output consists of a short non-gap buffer followed by a gap
buffer, the code was forgetting to flush the adapter for the gap buffer.
this caused an accumulation of data inside the element, and a growing shift
between the input and output streams, and resulted in pipeline lock-ups
(and, of course, totally nonsense \chi^{2} values, but we hadn't spotted
that problem yet).