remove default lib paths from PYTHON_LDFLAGS
this is a hack to remove the linker default search paths from -L options in
PYTHON_LDFLAGS.  this is needed to allow python modules to be built by
libtool when gstlal is already installed system wide.  I don't know where
the problem ultimately lies, but this makes it go away for now.