add framecpp_igwdparse element
this element implements a gstreamer parser for IGWD frame files.  i'm
pretty sure it's totally borken:  it only works for files in which 4-byte
ints are 4 bytes, etc., and I have no idea how to identify an end-of-file
marker robustly so I just copied the "class" code I found in the hex dump
of a frame file on my own computer.  anyway, it's enough to let me run
totem on a frame file and have it actually do something! :-).
unfortunately it just sits there because it thinks it needs to wait several
hundred million seconds before rendering something, but that'll get sorted
out eventually.

this work includes the addition of a "parsed=true" field to the caps
claimed by the framecpp_channeldemux and gds_lvshmsrc elements.  meanwhile
the igwd typefinder reports "parsed=false", which induces gstreamer's
autoplugger to insert the new parser element between generic source
elements and the channel demuxer.