gstlal.inspiral.py:  fix a "lock up"
if the user forgets to provide a marginalized likelihood file the program
attempts to read likelihood data from stdin, making it appear as though
the program has locked up.  the user could conceivably type likelihood data
in xml format into their terminal to get the program running again, but
this is unlikely to be a useful mode of operation.