framecpp:  add channelmux element
The element doesn't actually do anything yet.  Well, it does a *lot* but it
can't build frame files.  I'm just commiting it so that Ed and I can both
see where it's at and talk about how to get it to assemble a frame file.
It compiles, though, and gst-inspect seems happy with it, so have fun.