s5_pbh_trigger_pipe: add bin-by-mass-ratio to plot_sensitivity node
bin-by-mass-ratio is a new option for plot_sensitivity. This commits
makes that option get used by default in trigger_pipe. There is also
a new option bin-by-aligned-spin, but I'm not adding it now because I
think presently it will cause plot_sensitivity to crash if there are
no injections with spin and you ask it to bin in spin.