convert lvshmsrc to C++
Not really, just rename it and fix the warnings.  The motivation is to,
next, port it to the C++ lvshm API.  Why?  There isn't a C++ api for the
writer, so the matching sink element uses the C++ api creating an asymmetry
in the structure of this element.  Also, I think this is the only code in
the universe that uses the C api so (i) it's probably more prone to bugs
than the more widely-used C++ api and (ii) if we stop using it John could
stop supporting it and have one less thing to have to worry about.