gstlal-inspiral: implement autocorrelation mask property for filtering
For high mass systems, the template autocorrelation is a strong function
of the template duration. Consequently, the efficacy the autocorrelation
chisq test varies significantly across the high mass template space.
The autocorrelation mask allows one to have a template-dependent
autocorrelation length. In this commit, the autocorrelation mask is just
an array of ones. In the future, this mask needs to be tuned.