bank_splitter: remove references to sngl_inspiral.chi
We no longer rely on the chi column of sngl_inspiral to be
populated. NB: "chi" is different for different waveforms. Implemented
here is the effective spin usually associated with IMRPhenomB and
company. For the reduced spin chi (e.g. as in TaylorF2RedSpin), the
effective spin is not strictly correct but is probably close enough
for the purpose of splitting the bank and anyway is certainly better
than relying on a defunct sngl_inspiral column.