gstlal:  lib/ --> lib/gstlal/
this makes the paths to header files in #include directives the same for
code both in and out of tree.  this helps make it possible to subclass
gstlal elements in other packages since the element header files can now be
used out-of-tree, and it makes it eaiser to move code from other packages
into gstlal core (e.g. the framecpp plugin)