AppSync:  try to fix a race condition in exception reporting
if the call-back to handle a new buffer raises an exception, the lock can
get released in the finally: clause allowing other threads waiting to go to
begin running before the finally: clause re-raises the exception.  if the
thread that starts running hits the assert related to un-processed buffers
then you get that exception instead of the one that really triggered the
failure, making debugging difficult.  this patch moves the call-back to the
very end of the try: clause so that there's no book-keeping left to take
care of in the event that it fails, which should ensure we get the correct
traceback displayed.