excesspower.py: Try to align the filters with the PSD.
If the user doesn't specify a power of two for the search bandwidth
and/or filter bandwidth, then the whitener may complain that the
frequency series for either are not aligned and refuse to continue. This
attempts to circumvent this by realigning the filter by offsetting it
slightly. In most cases this is unlikely to work due to floating point
comparisons in LAL's AverageSpectrum.c and so error messages are made
even more verbose to explain this.