framexmitsrc:  don't update latency when timeouts occur
there's no right answer here, but when a timeout occurs the latency is used to guess the timestamp of the missing data, and if that timestamp is then used to udpate the latency the result is the latency keeps growing in step with the passage of time and the fake buffers timestamps all end up being identical.  that's arguably the right answer but not helpful.  by not recalculating the latency in the event of a timeout the fake buffers get monotonically increasing timestamps and the latency is left at whatever it was last measured to be before data stopped flowing.