Fixing channel-mask on framecpp_channeldemux
For some reason the function call

    caps = gst_caps_new_simple("audio/x-raw",
        "rate", G_TYPE_INT, *rate,
        "channels", G_TYPE_INT, 1,
        "layout", G_TYPE_STRING, "interleaved",
		"channel-mask", GST_TYPE_BITMASK, 0,
        NULL
    );

was leading to a channel-mask on the source pads of the demuxer
which actually had the value of a pointer instead of a channel mask.
This then would break downstream negotiation if the downstream
elements were not prepared to be flexible in the channel-mask value.
Oddly, by adding seemingly unrelated and trivial expressions above
this function call the issue would also be resolved, making me think
this is perhaps an error in the gstvalue.c code somewhere.

In any event, we can set it separately by hand afterward, as is done
here.

This element was tested in three ways:

First pipeline:
gst-launch-1.0 lal_cachesrc location="$HOME/data/test.cache" ! \
  progressreport update-freq=1 ! framecpp_channeldemux .H1:GDS-CALIB_STRAIN ! \
  taginject tags="instrument=H1,channel-name=GDS-CALIB_STRAIN,units=\"strain\""  \
  ! progressreport update-freq=1 ! fakesink

Second pipeline, ensuring that works with muxer:
gst-launch-1.0 lal_cachesrc location="$HOME/data/test.cache" ! \
  progressreport update-freq=1 ! framecpp_channeldemux .H1:GDS-CALIB_STRAIN ! \
  taginject tags="instrument=H1,channel-name=GDS-CALIB_STRAIN,units=\"strain\"" \
  ! framecpp_channelmux ! progressreport update-freq=1 ! fakesink

Third pipeline, ensuring works with injections:
gst-launch-1.0 lal_cachesrc location="$HOME/data/test.cache" ! \
  framecpp_channeldemux .H1:GDS-CALIB_STRAIN ! \
  taginject tags="instrument=H1,channel-name=GDS-CALIB_STRAIN,units=\"strain\"" ! \
  lal_simulation xml_location="$HOME/data/post-O1_bns_injs_jan10-jun10.xml" ! \
  framecpp_channelmux ! \
  framecpp_filesink async-handling=true path="$HOME/data" instrument="H1" frame_type="inj"

In this last pipeline we still end up with an error which has to be
attended to:
(gst-launch-1.0:2424756): GStreamer-CRITICAL **:
Trying to dispose object "frpad0", but it still has a parent "framecppchannelmux0".
You need to let the parent manage the object instead of unreffing the object directly.