Enable upload single-detector SNR time series to GraceDB
The old approach of storing a pointer to the COMPLEX8TimeSeries in
the GSTLALSnglInspiral struct did not work because we would often
add more than gst_buffer_get_max_memory() memory objects to a buffer.
Consequently, GStreamer would merge the memory objects together,
deallocating the original objects and causing the COMPLEX8TimeSeries
to be freed too early. There is no safe way to use a GstBuffer to
store a pointer to out-of-band auxiliary data.

The new approach is to store the SNR time series samples inline in
the buffer.