Add explicit synchronization point.
If there's a large number of parent / child relationships, instead
of having a set of all-to-all arcs (which can number into the tens
of millions for many parents and children in one line), add in a
explicit no-op node.  This causes the number of arcs to grow linearly
instead of quadratically.

HTCondor will do this automatically in future versions; this commit
should be reverted once the new dagman is available.