string_triggergen.c: fix gap treatment
Fixes a bug that when two triggers are sandwiching a gap, the pipeline regards as a single trigger with duration including the gap.
The commit fixes so that the pipeline will regard this as two separate triggers.
Also the durations of triggers are added by two samples, to be consistent with the previous pipeline.